# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TaskFlow Backend API",
    "version": "1.0.0",
    "description": "Multi-Tenant Project & Task Management REST API with Asynchronous Notification Queues and RBAC Security."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Provide valid JWT Access Token"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "organization_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "project_id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["todo", "in_progress", "review", "done"] },
          "priority": { "type": "string", "enum": ["low", "medium", "high", "urgent"] },
          "due_date": { "type": "string", "format": "date-time", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "JobResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "active", "completed", "failed"] },
          "metadata": {
            "type": "object",
            "properties": {
              "taskId": { "type": "string" },
              "taskTitle": { "type": "string" },
              "assignedUserId": { "type": "string" },
              "assignedUserEmail": { "type": "string" },
              "assignedByUserId": { "type": "string" },
              "attemptsMade": { "type": "integer" },
              "failedReason": { "type": "string", "nullable": true },
              "timestamp": { "type": "integer" }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": { "type": "object" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register new Organization Admin User",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "name", "password", "organizationName"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "name": { "type": "string" },
                  "password": { "type": "string", "minLength": 8 },
                  "organizationName": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Registration Successful" },
          "400": { "description": "Validation Error" },
          "409": { "description": "Email Already Exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login Authentication",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Authentication Successful" },
          "401": { "description": "Invalid Credentials" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "summary": "Rotate Access & Refresh Tokens",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": { "refreshToken": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tokens Rotated Successfully" },
          "401": { "description": "Invalid or Expired Refresh Token" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Revoke Refresh Token",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["refreshToken"],
                "properties": { "refreshToken": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Token Revoked Successfully" }
        }
      }
    },
    "/projects": {
      "post": {
        "summary": "Create Organization Project",
        "tags": ["Projects"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Project Created" },
          "401": { "description": "Unauthorized" }
        }
      },
      "get": {
        "summary": "List Organization Projects",
        "tags": ["Projects"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "string", "default": "1" } },
          { "name": "limit", "in": "query", "schema": { "type": "string", "default": "20" } }
        ],
        "responses": {
          "200": { "description": "Paginated List of Projects" }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "summary": "Get Project by ID",
        "tags": ["Projects"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Project Details" },
          "403": { "description": "Forbidden Cross-Tenant Access" },
          "404": { "description": "Project Not Found" }
        }
      },
      "patch": {
        "summary": "Update Project",
        "tags": ["Projects"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Project Updated" }
        }
      },
      "delete": {
        "summary": "Delete Project (Requires org_admin Role)",
        "tags": ["Projects"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Project Deleted" },
          "403": { "description": "Forbidden (Requires org_admin role)" }
        }
      }
    },
    "/projects/{id}/dashboard": {
      "get": {
        "summary": "Get Project Task Status Dashboard Summary",
        "tags": ["Projects"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task status breakdown counts (todo, in_progress, review, done)" }
        }
      }
    },
    "/projects/{projectId}/tasks": {
      "post": {
        "summary": "Create Task within Project",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["todo", "in_progress", "review", "done"] },
                  "priority": { "type": "string", "enum": ["low", "medium", "high", "urgent"] },
                  "dueDate": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task Created" }
        }
      },
      "get": {
        "summary": "List Project Tasks with Multi-Criteria Filtering",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["todo", "in_progress", "review", "done"] } },
          { "name": "priority", "in": "query", "schema": { "type": "string", "enum": ["low", "medium", "high", "urgent"] } },
          { "name": "assigneeId", "in": "query", "schema": { "type": "string" } },
          { "name": "dueFrom", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "dueTo", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "page", "in": "query", "schema": { "type": "string", "default": "1" } },
          { "name": "limit", "in": "query", "schema": { "type": "string", "default": "20" } }
        ],
        "responses": {
          "200": { "description": "Paginated & Filtered Tasks List" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get Task Details",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task Details" },
          "404": { "description": "Task Not Found" }
        }
      },
      "patch": {
        "summary": "Update Task",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["todo", "in_progress", "review", "done"] },
                  "priority": { "type": "string", "enum": ["low", "medium", "high", "urgent"] },
                  "dueDate": { "type": "string", "format": "date-time", "nullable": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task Updated" }
        }
      },
      "delete": {
        "summary": "Delete Task",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task Deleted" }
        }
      }
    },
    "/tasks/{id}/assignments": {
      "post": {
        "summary": "Assign User to Task & Enqueue Notification Job",
        "tags": ["Assignments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId"],
                "properties": { "userId": { "type": "string", "format": "uuid" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User Assigned & Job Enqueued (returns jobId)" },
          "403": { "description": "User does not belong to organization" },
          "409": { "description": "Duplicate Assignment" }
        }
      }
    },
    "/tasks/{id}/assignments/{userId}": {
      "delete": {
        "summary": "Unassign User from Task",
        "tags": ["Assignments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "User Unassigned" },
          "404": { "description": "Assignment Not Found" }
        }
      }
    },
    "/jobs/{id}": {
      "get": {
        "summary": "Get Notification Job Processing Status",
        "tags": ["Jobs"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Job Status (pending, active, completed, failed)" },
          "403": { "description": "Forbidden Cross-Tenant Access" },
          "404": { "description": "Job Not Found" }
        }
      }
    }
  }
}